move subquery expressions
into their own module
